Separate update_data_after contextmanager to pootle_stores
